+++ ADTs/Tree.c
@@ -6,12 +6,26 @@
 #include "RedBlackTree.h"
 #include "SplayTree.h"
 
-static const TreeOps SPL_ops = {
+static void* BST_Insert(void* tree, int val) {
+    return InsertNode((BinaryTreeADT) tree, NewTreeNode(val));
+}
+
+static TreeOps AVL_ops = {
+    .insert = (void* (*)(void*, int)) AVL_Insert,
+    .print = NULL,
+};
+
+static TreeOps BST_ops = {
+    .insert = (void* (*)(void*, int))BST_Insert,
+    .print = NULL,
+};
+
+static TreeOps SPL_ops = {
     .insert = (void* (*)(void*, int))Splay_Insert,
     .print = NULL,
 };
 
-static const TreeOps RBT_ops = {
+static TreeOps RBT_ops = {
     .insert = (void* (*)(void*, int))Red_Black_Insert,
     .print = NULL,
 };
@@ -24,8 +38,10 @@ TreePtr tree_init(TreeType type) {
     assert(TREE_TYPE_NUMBER == 4 && "HAVE NOT EXHAUST ALL TREE TYPES");
     switch (type) {
         case AVL:
+            tree->ops = &AVL_ops;
             break;
         case BST:
+            tree->ops = &BST_ops;
             break;
         case SPL:
             tree->ops = &SPL_ops;